+++ game/res/ui.go
@@ -47,7 +47,7 @@ const helpText = "Tiny World Help" +
 	" - Pan: Arrows, WASD or middle mouse button\n" +
 	" - Zoom: +/- or mouse wheel\n" +
 	" - Pause/resume: Space\n" +
-	" - Game speed: PageUp / PageDown\n" +
+	" - Game speed: [/] (square brackets)\n" +
 	" - Toggle fullscreen: F11"
 
 const helpPanelWidth = 680
@@ -734,7 +734,7 @@ func (ui *UI) createInfo() *widget.Container {
 	infoContainer.AddChild(cont)
 	ui.timerLabel = lab
 
-	cont, lab = ui.createLabel("", "Game speed.\nControl with PageUp/PageDown/Space.", 35, widget.TextPositionEnd)
+	cont, lab = ui.createLabel("", "Game speed.\nControl with [/] (square brackets) and Space.", 35, widget.TextPositionEnd)
 	infoContainer.AddChild(cont)
 	ui.speedLabel = lab